reafactor: move nodes and ports in to their own file
